+++ pkg/workers/api/handlers.go
@@ -1,26 +1,26 @@
 package api
 
 import (
-	"context"
-	"encoding/base64"
+	"fmt"
+	"net/http"
 	"strconv"
-	"strings"
 	"text/template"
 	"time"
 
+	"github.com/allenta/varnishmon/pkg/config"
 	"github.com/allenta/varnishmon/pkg/workers/storage"
-	"github.com/fasthttp/router"
+	"github.com/gin-contrib/gzip"
+	"github.com/gin-contrib/pprof"
+	"github.com/gin-gonic/gin"
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/client_golang/prometheus/promhttp"
-	"github.com/valyala/fasthttp"
-	"github.com/valyala/fasthttp/fasthttpadaptor"
-	"github.com/valyala/fasthttp/pprofhandler"
 )
 
 type Handler struct {
-	app     Application
-	storage *storage.Storage
-	router  *router.Router
+	app          Application
+	serverHeader string
+	storage      *storage.Storage
+	router       *gin.Engine
 
 	homeTemplate *template.Template
 
@@ -30,10 +30,13 @@ type Handler struct {
 }
 
 func NewHandler(app Application, storage *storage.Storage) *Handler {
+	gin.SetMode(gin.ReleaseMode)
+
 	h := &Handler{
-		app:     app,
-		storage: storage,
-		router:  router.New(),
+		app:          app,
+		serverHeader: fmt.Sprintf("varnishmon/%s (%s)", config.Version(), config.Revision()),
+		storage:      storage,
+		router:       gin.New(),
 
 		requestsTotal: prometheus.NewCounterVec(
 			prometheus.CounterOpts{
@@ -59,13 +62,39 @@ func NewHandler(app Application, storage *storage.Storage) *Handler {
 	}
 
 	h.router.RedirectTrailingSlash = true
-	h.router.GET("/debug/pprof/{profile:*}", h.handlePprofRequest)
+	h.router.RedirectFixedPath = false
+	h.router.HandleMethodNotAllowed = false
+	h.router.ForwardedByClientIP = true
+	h.router.UseRawPath = false
+	h.router.UnescapePathValues = true
+	h.router.SetTrustedProxies(nil) //nolint:errcheck
+
+	h.router.Use(
+		gin.CustomRecoveryWithWriter(
+			h.app.Cfg().Log().ErrorWriter(),
+			func(c *gin.Context, recovered any) {
+				h.app.Cfg().Log().Error().
+					Interface("recovered", recovered).
+					Str("url", c.Request.URL.String()).
+					Str("method", c.Request.Method).
+					Msg("Recovered from gin panic!")
+				c.AbortWithStatus(http.StatusInternalServerError)
+			}),
+		gzip.Gzip(gzip.DefaultCompression, gzip.WithDecompressFn(gzip.DefaultDecompressHandle)))
+	if app.Cfg().APIBasicAuthUsername() != "" && app.Cfg().APIBasicAuthPassword() != "" {
+		h.router.Use(
+			gin.BasicAuth(gin.Accounts{
+				app.Cfg().APIBasicAuthUsername(): app.Cfg().APIBasicAuthPassword(),
+			}))
+	}
+
+	pprof.Register(h.router)
 	h.router.GET("/metrics", h.handleMetricsRequest)
 	h.router.GET("/storage/metrics", h.handleStorageMetricsRequest)
-	h.router.GET("/storage/metrics/{id:[0-9]+}", h.handleStorageMetricsRequest)
+	h.router.GET("/storage/metrics/:id", h.handleStorageMetricsRequest)
 	h.router.GET("/", h.handleHomeRequest)
 	h.router.GET("/config", h.handleConfigRequest)
-	h.router.ServeFilesCustom("/{filepath:*}", h.filesystemHandler())
+	h.router.NoRoute(h.filesystemHandler())
 
 	h.app.Cfg().Metrics().Registry.MustRegister(h.requestsTotal)
 	h.app.Cfg().Metrics().Registry.MustRegister(h.requestsInflightTotal)
@@ -74,57 +103,50 @@ func NewHandler(app Application, storage *storage.Storage) *Handler {
 	return h
 }
 
-func (h *Handler) HandleRequest(_ context.Context, rctx *fasthttp.RequestCtx) {
+type responseWriter struct {
+	http.ResponseWriter
+	code int
+}
+
+func (rw *responseWriter) WriteHeader(code int) {
+	rw.code = code
+	rw.ResponseWriter.WriteHeader(code)
+}
+
+func (h *Handler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
+	// Gin skips the execution of middlewares in some cases (e.g., automatic
+	// redirects due to trailing slashes). This is not a major issue, but it is
+	// the reason why the following logic is here and not part of a middleware.
+
+	// Wrap the original 'ResponseWriter' in order to capture the status code,
+	// needed for metrics.
+	rw := &responseWriter{ResponseWriter: w}
+
 	// Update metrics.
 	h.requestsInflightTotal.Inc()
 	defer func(method string, start time.Time) {
 		elapsed := time.Since(start).Seconds()
-		code := strconv.Itoa(rctx.Response.StatusCode())
+		code := strconv.Itoa(rw.code)
 
 		h.requestsTotal.WithLabelValues(method, code).Inc()
 		h.requestsInflightTotal.Dec()
 		h.requestDurationSeconds.WithLabelValues(method, code).Observe(elapsed)
-	}(string(rctx.Method()), time.Now())
+	}(r.Method, time.Now())
 
-	// Set no-cache headers for all responses.
-	rctx.Response.Header.Set("Cache-Control", "no-cache, no-store, must-revalidate")
-	rctx.Response.Header.Set("Pragma", "no-cache")
-	rctx.Response.Header.Set("Expires", "0")
+	// Set default 'Server' header for all responses.
+	rw.Header().Set("Server", h.serverHeader)
 
-	// Check authentication.
-	if h.app.Cfg().APIBasicAuthUsername() != "" && h.app.Cfg().APIBasicAuthPassword() != "" {
-		authorized := false
+	// Set default no-cache headers for all responses.
+	rw.Header().Set("Cache-Control", "no-cache, no-store, must-revalidate")
+	rw.Header().Set("Pragma", "no-cache")
+	rw.Header().Set("Expires", "0")
 
-		const prefix = "Basic "
-		authHeader := string(rctx.Request.Header.Peek("Authorization"))
-		if strings.HasPrefix(authHeader, prefix) {
-			if c, err := base64.StdEncoding.DecodeString(authHeader[len(prefix):]); err == nil {
-				cs := string(c)
-				s := strings.IndexByte(cs, ':')
-				authorized = s >= 0 &&
-					h.app.Cfg().APIBasicAuthUsername() == cs[:s] &&
-					h.app.Cfg().APIBasicAuthPassword() == cs[s+1:]
-			}
-		}
-
-		if !authorized {
-			rctx.SetStatusCode(fasthttp.StatusUnauthorized)
-			rctx.Response.Header.Add("WWW-Authenticate", "Basic realm=Restricted")
-			return
-		}
-	}
-
-	// Route request.
-	h.router.Handler(rctx)
+	// Route the request.
+	h.router.ServeHTTP(rw, r)
 }
 
-func (h *Handler) handlePprofRequest(rctx *fasthttp.RequestCtx) {
-	pprofhandler.PprofHandler(rctx)
-}
-
-func (h *Handler) handleMetricsRequest(rctx *fasthttp.RequestCtx) {
-	handler := fasthttpadaptor.NewFastHTTPHandler(promhttp.HandlerFor(
+func (h *Handler) handleMetricsRequest(c *gin.Context) {
+	promhttp.HandlerFor(
 		h.app.Cfg().Metrics().Registry,
-		promhttp.HandlerOpts{}))
-	handler(rctx)
+		promhttp.HandlerOpts{}).ServeHTTP(c.Writer, c.Request)
 }