+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+- ?
+    + Dropped fasthttp in favor of the standard net/http package.
+    + Integrated gin for routing.
+    + Dropped 'api_worker_concurrency' & 'api_worker_open_connections' metrics.
+
 - 0.6.6-1 (2025-03-10):
     + Added `/config` endpoint to retrieve the configuration, as an alternative to hydrating the `index.html` template server-side.